% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotreer-package.R
\docType{package}
\name{annotreer-package}
\alias{annotreer}
\alias{annotreer-package}
\title{annotreer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jtr13/annotreer}
  \item Report bugs at \url{https://github.com/jtr13/annotreer/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
