% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotree.R
\name{annotree}
\alias{annotree}
\title{Create Annotated Labels for rpart Trees}
\usage{
annotree(
  mod,
  show_classes = TRUE,
  show_gt = FALSE,
  show_cp = FALSE,
  show_pruned = FALSE,
  show_min = FALSE,
  show_leaf_dev = FALSE,
  show_internal_dev = FALSE,
  main = NULL
)
}
\arguments{
\item{mod}{an \code{rpart} object}

\item{show_classes}{logical. If \code{TRUE} show class labels and counts. Defaults to \code{TRUE}.}

\item{show_gt}{logical. If \code{TRUE} the g(t) term (misclassification reduction per split) is included. Defaults to \code{FALSE}.}

\item{show_cp}{logical. If \code{TRUE} the complexity parameter (cp) value, calculated as g(t) divided by root node deviance is included. Defaults to \code{FALSE}.}

\item{show_pruned}{logical. If \code{TRUE} the nodes slated for pruning are plotted in a distinct color. Defaults to \code{FALSE}.}

\item{show_min}{logical. If \code{TRUE} the node with minimum g(t) is labeled "*minimum". Defaults to \code{FALSE}.}

\item{show_leaf_dev}{logical. If \code{TRUE} leaves are labeled with "dev" (misclassified cases). Defaults to \code{FALSE}.}

\item{show_internal_dev}{logical. If \code{TRUE} internal nodes are labeled with "dev" (misclassified cases). Defaults to \code{FALSE}.}

\item{main}{An optional title for the plot.}
}
\description{
Generates custom labels for each node in an \code{rpart} classification tree to aid
in understanding CART's pruning algorithm
}
\details{
This function is primarily used to create informative labels for internal nodes and leaves in classification trees.
\itemize{
\item For leaf nodes, it displays the word \code{"leaf"} and the node's deviance.
\item For internal nodes, it can display the g(t) equation, cp value, and optionally indicate the node with the smallest g(t).
}
}
\examples{
library(rpart)
kmod <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis, cp = 0, minsplit = 2)
annotree(kmod)

}
